penaltyslab

+++ MangalWeb.Model/Entity/Mst_PenaltySlab.cs
@@ -16,7 +16,7 @@ namespace MangalWeb.Model.Entity
     {
         public int Ps_Id { get; set; }
         public System.DateTime Ps_Datewef { get; set; }
-        public decimal Ps_Penalty { get; set; }
+        public double Ps_Penalty { get; set; }
         public int Ps_Accounthead { get; set; }
         public Nullable<System.DateTime> Ps_RecordCreated { get; set; }
         public Nullable<System.DateTime> Ps_RecordUpdated { get; set; }
